backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.0.1
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.3.0
openai>=1.30.0
python-multipart==0.0.6
gunicorn==21.2.0
whitenoise==6.6.0